# exam-gu-server/api/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: support@examgu.uqac.ca
    name: Support Exam-GU
  description: API de gestion des examens en ligne pour l'Université de Québec à Chicoutimi
  title: Exam-GU API
  version: 1.0.0
servers:
- description: Serveur de développement
  url: http://localhost:8080/api/v1
- description: Serveur de production
  url: https://api.examgu.uqac.ca/api/v1
security:
- BearerAuth: []
...
paths:
  /auth/login:
    post:
      description: Authentifier un utilisateur et obtenir un token JWT
      operationId: authLoginPOST
      x-eov-operation-handler: controllers/AuthenticationController
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/_auth_login_post_request"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_auth_login_post_200_response"
          description: Authentification réussie
        "401":
          description: Identifiants invalides
      summary: Se connecter
      tags:
      - Authentication
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: Authentication
...
  /teacher/exams:
    get:
      summary: Mes examens
      description: Retourne tous les examens créés par l'enseignant
      operationId: teacherExamsGET
      x-eov-operation-handler: controllers/TeacherExamController
      tags:
        - Teacher
      responses:
        "200":
          description: Liste des examens
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
        "401":
          description: Non authentifié
    post:
      summary: Créer un examen
      description: Créer un nouvel examen pour une session
      operationId: teacherExamsPOST
      x-eov-operation-handler: controllers/TeacherExamController
      tags:
        - Teacher
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - sessionId
                - title
                - duration
                - startTime
                - endTime
              properties:
                sessionId:
                  type: string
                title:
                  type: string
                description:
                  type: string
                duration:
                  type: integer
                  description: Durée en minutes
                startTime:
                  type: string
                  format: date-time
                endTime:
                  type: string
                  format: date-time
                maxAttempts:
                  type: integer
                  default: 1
      responses:
        "201":
          description: Examen créé
        "400":
          description: Données invalides
        "401":
          description: Non authentifié
        "403":
          description: Accès refusé
  /teacher/exams/{examId}:
    get:
      summary: Détails d'un examen
      description: Récupérer les détails complets d'un examen
      operationId: teacherExamsExamIdGET
      x-eov-operation-handler: controllers/TeacherExamController
      tags:
        - Teacher
      parameters:
        - name: examId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Détails de l'examen
        "401":
          description: Non authentifié
        "403":
          description: Accès refusé
        "404":
          description: Examen non trouvé
    put:
      summary: Modifier un examen
      description: Modifier les informations d'un examen
      operationId: teacherExamsExamIdPUT
      x-eov-operation-handler: controllers/TeacherExamController
      tags:
        - Teacher
      parameters:
        - name: examId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                duration:
                  type: integer
                startTime:
                  type: string
                  format: date-time
                endTime:
                  type: string
                  format: date-time
                maxAttempts:
                  type: integer
      responses:
        "200":
          description: Examen modifié
        "400":
          description: Modification impossible
        "401":
          description: Non authentifié
        "403":
          description: Accès refusé
        "404":
          description: Examen non trouvé
    delete:
      summary: Supprimer un examen
      description: Supprimer un examen (impossible s'il y a des tentatives)
      operationId: teacherExamsExamIdDELETE
      x-eov-operation-handler: controllers/TeacherExamController
      tags:
        - Teacher
      parameters:
        - name: examId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Examen supprimé
        "400":
          description: Suppression impossible
        "401":
          description: Non authentifié
        "403":
          description: Accès refusé
        "404":
          description: Examen non trouvé
...
  /student/exams:
    get:
      summary: Mes examens disponibles
      description: Retourne tous les examens disponibles pour l'étudiant
      operationId: studentExamsGET
      x-eov-operation-handler: controllers/StudentExamController
      tags:
        - Student
      responses:
        "200":
          description: Liste des examens
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
        "401":
          description: Non authentifié
  /student/exams/{examId}:
    get:
      summary: Détails d'un examen
      description: Récupérer les détails d'un examen (sans les réponses)
      operationId: studentExamsExamIdGET
      x-eov-operation-handler: controllers/StudentExamController
      tags:
        - Student
      parameters:
        - name: examId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Détails de l'examen
        "401":
          description: Non authentifié
        "403":
          description: Accès refusé
        "404":
          description: Examen non trouvé
...
  /admin/users:
    get:
      description: Retourne la liste de tous les utilisateurs (Admin seulement)
      operationId: adminUsersGET
      x-eov-operation-handler: controllers/AdminController
      parameters:
      - description: Filtrer par rôle
        explode: true
        in: query
        name: role
        required: false
        schema:
          enum:
          - ADMIN
          - TEACHER
          - STUDENT
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/User"
                type: array
          description: Liste des utilisateurs
        "403":
          description: Accès refusé (Admin requis)
      summary: Lister les utilisateurs
      tags:
      - Admin
      x-accepts:
      - application/json
      x-tags:
      - tag: Admin
    post:
      description: Créer un nouvel enseignant ou étudiant (Admin seulement)
      operationId: adminUsersPOST
      x-eov-operation-handler: controllers/AdminController
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
          description: Utilisateur créé
        "400":
          description: Données invalides
        "403":
          description: Accès refusé (Admin requis)
      summary: Créer un utilisateur
      tags:
      - Admin
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: Admin
  /admin/users/{userId}:
    delete:
      operationId: adminUsersUserIdDELETE
      x-eov-operation-handler: controllers/AdminController
      parameters:
      - explode: false
        in: path
        name: userId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: Utilisateur supprimé
        "404":
          description: Utilisateur non trouvé
        "403":
          description: Accès refusé
      summary: Supprimer un utilisateur
      tags:
      - Admin
      x-accepts:
      - application/json
      x-tags:
      - tag: Admin
      
    get:
      operationId: adminUsersUserIdGET
      x-eov-operation-handler: controllers/AdminController
      parameters:
      - explode: false
        in: path
        name: userId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
          description: Détails de l'utilisateur
        "404":
          description: Utilisateur non trouvé
        "403":
          description: Accès refusé
      summary: Obtenir un utilisateur
      tags:
      - Admin
      x-accepts:
      - application/json
      x-tags:
      - tag: Admin
      
    put:
      operationId: adminUsersUserIdPUT
      x-eov-operation-handler: controllers/AdminController
      parameters:
      - explode: false
        in: path
        name: userId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
          description: Utilisateur modifié
        "404":
          description: Utilisateur non trouvé
        "403":
          description: Accès refusé
      summary: Modifier un utilisateur
      tags:
      - Admin
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: Admin
...
  /exams:
    get:
      description: "Retourne les examens de l'utilisateur (ses examens pour enseignant,\
        \ ses examens inscrits pour étudiant)"
      operationId: examsGET
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/Exam"
                type: array
          description: Liste des examens
        "401":
          description: Non authentifié
      summary: Lister les examens
      tags:
      - Exams
      x-accepts:
      - application/json
      x-tags:
      - tag: Exams
      x-eov-operation-handler: controllers/ExamsController
    post:
      description: Créer un nouvel examen (Enseignant seulement)
      operationId: examsPOST
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExamRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exam"
          description: Examen créé
        "400":
          description: Données invalides
        "403":
          description: Accès refusé (Enseignant requis)
      summary: Créer un examen
      tags:
      - Exams
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: Exams
      x-eov-operation-handler: controllers/ExamsController
  /exams/{examId}:
    delete:
      description: Supprimer un examen en brouillon (Enseignant seulement)
      operationId: examsExamIdDELETE
      parameters:
      - explode: false
        in: path
        name: examId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: Examen supprimé
        "404":
          description: Examen non trouvé
        "403":
          description: Accès refusé ou examen non en brouillon
      summary: Supprimer un examen
      tags:
      - Exams
      x-accepts:
      - application/json
      x-tags:
      - tag: Exams
      x-eov-operation-handler: controllers/ExamsController
    get:
      description: Retourne les détails d'un examen si l'utilisateur a accès
      operationId: examsExamIdGET
      parameters:
      - explode: false
        in: path
        name: examId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamDetail"
          description: Détails de l'examen
        "404":
          description: Examen non trouvé
        "403":
          description: Accès refusé
      summary: Obtenir un examen
      tags:
      - Exams
      x-accepts:
      - application/json
      x-tags:
      - tag: Exams
      x-eov-operation-handler: controllers/ExamsController
    put:
      description: Modifier un examen en brouillon (Enseignant seulement)
      operationId: examsExamIdPUT
      parameters:
      - explode: false
        in: path
        name: examId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateExamRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exam"
          description: Examen modifié
        "404":
          description: Examen non trouvé
        "403":
          description: Accès refusé ou examen non en brouillon
      summary: Modifier un examen
      tags:
      - Exams
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: Exams
      x-eov-operation-handler: controllers/ExamsController
...
components:
  schemas:
    User:
      example:
        firstName: firstName
        lastName: lastName
        createdAt: 2000-01-23T04:56:07.000+00:00
        role: ADMIN
        id: id
        email: email
      properties:
        id:
          description: ID unique de l'utilisateur
          type: string
        email:
          format: email
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          enum:
          - ADMIN
          - TEACHER
          - STUDENT
          type: string
        createdAt:
          format: date-time
          type: string
      required:
      - email
      - firstName
      - id
      - lastName
      - role
      type: object
    CreateUserRequest:
      example:
        firstName: firstName
        lastName: lastName
        password: password
        role: TEACHER
        email: email
      properties:
        email:
          format: email
          type: string
        password:
          format: password
          minLength: 8
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          enum:
          - TEACHER
          - STUDENT
          type: string
      required:
      - email
      - firstName
      - lastName
      - password
      - role
      type: object
    UpdateUserRequest:
      example:
        firstName: firstName
        lastName: lastName
        email: email
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          format: email
          type: string
      type: object
...
    Exam:
      example:
        description: description
        sessionId: sessionId
        title: title
        duration: 0
        maxAttempts: 6
        createdAt: 2000-01-23T04:56:07.000+00:00
        teacherId: teacherId
        startTime: 2000-01-23T04:56:07.000+00:00
        id: id
        endTime: 2000-01-23T04:56:07.000+00:00
        courseId: courseId
        status: DRAFT
        updatedAt: 2000-01-23T04:56:07.000+00:00
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        sessionId:
          type: string
        courseId:
          type: string
        teacherId:
          type: string
        duration:
          description: Durée en minutes
          type: integer
        startTime:
          format: date-time
          type: string
        endTime:
          format: date-time
          type: string
        maxAttempts:
          description: Nombre maximum de tentatives
          type: integer
        status:
          enum:
          - DRAFT
          - ACTIVE
          - CLOSED
          - IN_GRADING
          - GRADED
          type: string
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
      required:
      - courseId
      - duration
      - id
      - maxAttempts
      - sessionId
      - status
      - teacherId
      - title
      type: object
    ExamDetail:
      allOf:
      - $ref: "#/components/schemas/Exam"
      - properties:
          questions:
            items:
              $ref: "#/components/schemas/Question"
            type: array
          questionCount:
            type: integer
        type: object
      example:
        questionCount: 5
        questions:
        - examId: examId
          id: id
          text: text
          type: MCQ
          points: 1.4658129805029452
          order: 5
        - examId: examId
          id: id
          text: text
          type: MCQ
          points: 1.4658129805029452
          order: 5
        description: description
        sessionId: sessionId
        title: title
        duration: 0
        maxAttempts: 6
        createdAt: 2000-01-23T04:56:07.000+00:00
        teacherId: teacherId
        startTime: 2000-01-23T04:56:07.000+00:00
        id: id
        endTime: 2000-01-23T04:56:07.000+00:00
        courseId: courseId
        status: DRAFT
        updatedAt: 2000-01-23T04:56:07.000+00:00
    CreateExamRequest:
      example:
        duration: 0
        maxAttempts: 1
        description: description
        startTime: 2000-01-23T04:56:07.000+00:00
        sessionId: sessionId
        endTime: 2000-01-23T04:56:07.000+00:00
        title: title
      properties:
        title:
          type: string
        description:
          type: string
        sessionId:
          type: string
        duration:
          description: Durée en minutes
          type: integer
        startTime:
          format: date-time
          type: string
        endTime:
          format: date-time
          type: string
        maxAttempts:
          minimum: 1
          type: integer
      required:
      - duration
      - endTime
      - maxAttempts
      - sessionId
      - startTime
      - title
      type: object
    UpdateExamRequest:
      example:
        duration: 0
        maxAttempts: 6
        description: description
        startTime: 2000-01-23T04:56:07.000+00:00
        endTime: 2000-01-23T04:56:07.000+00:00
        title: title
      properties:
        title:
          type: string
        description:
          type: string
        duration:
          type: integer
        startTime:
          format: date-time
          type: string
        endTime:
          format: date-time
          type: string
        maxAttempts:
          type: integer
      type: object
    Question:
      discriminator:
        propertyName: type
      example:
        examId: examId
        id: id
        text: text
        type: MCQ
        points: 1.4658129805029452
        order: 5
      oneOf:
      - $ref: "#/components/schemas/MCQQuestion"
      - $ref: "#/components/schemas/TrueFalseQuestion"
      - $ref: "#/components/schemas/OpenEndedQuestion"
      properties:
        id:
          type: string
        examId:
          type: string
        type:
          enum:
          - MCQ
          - TRUE_FALSE
          - OPEN_ENDED
          type: string
        text:
          type: string
        points:
          description: Points pour cette question
          type: number
        order:
          description: Ordre d'apparition dans l'examen
          type: integer
      required:
      - examId
      - id
      - points
      - text
      - type
      type: object
      x-one-of-name: Question
    MCQQuestion:
      properties:
        type:
          enum:
          - MCQ
          type: string
        options:
          items:
            $ref: "#/components/schemas/MCQOption"
          minItems: 2
          type: array
      required:
      - options
      type: object
    MCQOption:
      properties:
        id:
          type: string
        text:
          type: string
        isCorrect:
          type: boolean
      required:
      - id
      - isCorrect
      - text
      type: object
    TrueFalseQuestion:
      properties:
        type:
          enum:
          - TRUE_FALSE
          type: string
        correctAnswer:
          type: boolean
      required:
      - correctAnswer
      type: object
    OpenEndedQuestion:
      properties:
        type:
          enum:
          - OPEN_ENDED
          type: string
      type: object
...
    _auth_login_post_request:
      properties:
        email:
          example: enseignant@uqac.ca
          format: email
          type: string
        password:
          example: SecurePassword123
          format: password
          type: string
      required:
      - email
      - password
      type: object
    _auth_login_post_200_response:
      example:
        user:
          firstName: firstName
          lastName: lastName
          createdAt: 2000-01-23T04:56:07.000+00:00
          role: ADMIN
          id: id
          email: email
        token: token
      properties:
        token:
          description: Token JWT pour les requêtes futures
          type: string
        user:
          $ref: "#/components/schemas/User"
      type: object
...
  securitySchemes:
    BearerAuth:
      bearerFormat: JWT
      description: Authentification par token JWT
      scheme: bearer
      type: http